Resend with NACK ok and Recovering from buffer on receiver also when receiving missing packet avec having received the next ones

+++ src/packet/packet.c
@@ -232,7 +232,7 @@ pkt_status_code pkt_encode(pkt_t *pkt, char *buf, size_t *len) {
         //return E_UNCONSISTENT;
     }
 
-    *len = charWritten; // indiquer le nombre de chars ecrits
+    *len = (size_t) charWritten; // indiquer le nombre de chars ecrits
 
     return PKT_OK;
 }
